02-24-2009 Giansimon Diblas: Bug fixing.
git-svn-id: file:///home/wik/repos/w3srepo/sfW3studioCmsPlugin/1.2/trunk@63 c6a47aa4-7d28-49a2-9282-aebedf263cad

+++ modules/menuBuilder/lib/BaseW3sMenuBuilderActions.class.php
@@ -22,31 +22,56 @@ class BaseW3sMenuBuilderActions extends BaseW3sContentsManagerActions
 	 */
 	public function executeEdit($request)
   {     
-    $currentContent = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
-    if($this->getRequestParameter('setClassToPages') == 0)
-    { 
-	    $content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
-	    $result = $content->edit(array('Content' => $this->getRequestParameter('content')));
+    if ($request->hasParameter('idContent') && $request->hasParameter('content'))
+    {
+      $content = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
+      if (is_object($content))
+      {
+        if ($content->getContentTypeId() == 5)
+        {
+          $currentContent = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
+          if(!$request->hasParameter('setClassToPages') || $this->getRequestParameter('setClassToPages') == 0)
+          {
+            $content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
+            $result = $content->edit(array('Content' => $this->getRequestParameter('content')));
+          }
+          else
+          {
+            $content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
+            $menu = new w3sGroupForeignContentsMenu($content);
+            $result = $menu->update($this->getRequestParameter('content'));
+          }
+
+          if ($result == 1)
+          {
+            $currentContent = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
+            $content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
+
+            return $this->renderText($content->getDisplayContentForEditorMode());
+          }
+          else
+          {
+            $this->getResponse()->setStatusCode(404);
+            return $this->renderText($content->displayError($result, true));
+          }
+        }
+        else
+        {
+          $this->getResponse()->setStatusCode(404);
+          return $this->renderText('The content you tried to edit is not a valid navigation menu.');
+        }
+      }
+      else
+      {
+        $this->getResponse()->setStatusCode(404);
+        return $this->renderText('The content you tried to edit doesn\'t exist anymore.');
+      }
     }
     else
-    { 
-    	$content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
-    	$menu = new w3sGroupForeignContentsMenu($content);
-    	$result = $menu->update($this->getRequestParameter('content'));
+    {
+      $this->getResponse()->setStatusCode(404);
+      return $this->renderText('A required parameter misses.');
     }
-    
-    if ($result == 1)
-    { 
-    	$currentContent = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
-    	$content = w3sContentManagerFactory::create($currentContent->getContentTypeId(), $currentContent);
-    	
-    	return $this->renderText($content->getDisplayContentForEditorMode());
-    }
-    else
-    {     	
-    	$this->getResponse()->setStatusCode(404);
-    	return $this->renderText($content->displayError($result, true));
-    }    
   }
   
   /**
@@ -74,17 +99,33 @@ class BaseW3sMenuBuilderActions extends BaseW3sContentsManagerActions
   { 
     if ($request->hasParameter('idContent'))
     {
-      $this->menuEditor = new w3sMenuEditor(DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent')), 5);
-      //$this->menuEditor->saveLinks($this->getRequest()->getParameterHolder()->getAll());
-      $this->result = $this->menuEditor->saveMenuLink($this->getContext()->getI18N()->__('This is a link'));
-
-      if ($this->result != 0)
+      $content = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
+      if (is_object($content))
       {
-        $this->clearCache($this->getRequestParameter('idContent'));
+        if ($content->getContentTypeId() == 5)
+        {
+          $this->menuEditor = new w3sMenuEditor($content);
+          $this->result = $this->menuEditor->saveMenuLink($this->getContext()->getI18N()->__('This is a link'));
+
+          if ($this->result != 0)
+          {
+            $this->clearCache($this->getRequestParameter('idContent'));
+          }
+          else
+          {
+            $this->getResponse()->setStatusCode(404);
+          }
+        }
+        else
+        {
+          $this->getResponse()->setStatusCode(404);
+          return $this->renderText('The content you tried to edit is not a valid navigation menu.');
+        }
       }
       else
       {
         $this->getResponse()->setStatusCode(404);
+        return $this->renderText('The content you tried to edit doesn\'t exist anymore.');
       }
     }
     else
@@ -101,26 +142,42 @@ class BaseW3sMenuBuilderActions extends BaseW3sContentsManagerActions
   {
     if ($request->hasParameter('idContent') && $this->getRequestParameter('idMenu'))
     {
-      $this->menuEditor = new w3sMenuEditor(DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent')), 5);
-      
-      $menu = DbFinder::from('W3sMenuElement')->findPK($this->getRequestParameter('idMenu'));
-      $position = $menu->getPosition();
-      $menu->delete();
-      
-      $menuElements =  DbFinder::from('W3sMenuElement')->
-                       where('contentId', $this->getRequestParameter('idContent'))->
-                       where('Position', '>', $position)->
-                       orderBy('Position', 'ASC')->
-                       find(); print_R($menuElements);
-      foreach($menuElements as $menuElement)
+      $content = DbFinder::from('W3sContent')->findPK($this->getRequestParameter('idContent'));
+      if (is_object($content))
       {
-        echo $position."\n";
-        $menuElement->setPosition($position);
-        $menuElement->save();
-        $position++;
-      }
+        if ($content->getContentTypeId() == 5)
+        {
+          $this->menuEditor = new w3sMenuEditor($content);
 
-      $this->clearCache($this->getRequestParameter('idContent'));
+          $menu = DbFinder::from('W3sMenuElement')->findPK($this->getRequestParameter('idMenu'));
+          $position = $menu->getPosition();
+          $menu->delete();
+
+          $menuElements =  DbFinder::from('W3sMenuElement')->
+                           where('contentId', $this->getRequestParameter('idContent'))->
+                           where('Position', '>', $position)->
+                           orderBy('Position', 'ASC')->
+                           find(); print_R($menuElements);
+          foreach($menuElements as $menuElement)
+          {
+            $menuElement->setPosition($position);
+            $menuElement->save();
+            $position++;
+          }
+
+          $this->clearCache($this->getRequestParameter('idContent'));
+        }
+        else
+        {
+          $this->getResponse()->setStatusCode(404);
+          return $this->renderText('The content you tried to delete is not a valid navigation menu.');
+        }
+      }
+      else
+      {
+        $this->getResponse()->setStatusCode(404);
+        return $this->renderText('The content you tried to edit doesn\'t exist anymore.');
+      }
     }
     else
     {
@@ -130,101 +187,32 @@ class BaseW3sMenuBuilderActions extends BaseW3sContentsManagerActions
   }
   
   /**
-	 * Deletes a link from the current menu
-	 */
-  public function executeChangePosition()
+	 * Deletes a link from the current menu. Obsolete
+	 
+  public function executeChangePosition($request)
   {
-    $i = 1;
-    $menuList = $this->getRequestParameter('w3s_mb_list');
-    foreach($menuList as $idMenu)
+    if ($request->hasParameter('w3s_mb_list'))
     {
-    	$menu = DbFinder::from('W3sMenuElement')->findPK($idMenu);
-			$menu->setPosition($i);
-			$menu->save();
-			$i++;
-    }
-    
-    return sfView::NONE;
-  }
-
-/*
-  public function executeSave()
-  {
-    // We assure that all the operations W3StudioCMS makes will be successfully done
-    $con = Propel::getConnection();
-    $bRollBack = false;
-    $con->begin();
-    
-    $this->saveLinks($this->getRequest()->getParameterHolder()->getAll());
-
-    $oLanguage = W3sLanguagePeer::retrieveByPK($this->getRequestParameter('lang'));
-    $language = $oLanguage->getLanguage();
-    $menu = W3sMenuElementPeer::getContentMenu($this->getRequestParameter('idContent'));
-
-    if($this->getRequestParameter('setClassToPages') == 1){
-      $oMenuElement = new W3sMenuElement();
-      $group = W3sGroupPeer::retrieveByPk($this->getRequestParameter('idGroup'));
-      foreach($group->getW3sPages() as $page):
-        $idPage = $page->getId();
-        $currentContents = W3sContentPeer::getSlotContents($this->getRequestParameter('idSlot'), $this->getRequestParameter('lang'), $idPage);       
-        foreach($currentContents as $currentContent):
-        	//echo $oPage->getPageName() . ' - ' . $this->getRequestParameter('classAssigned') . '<br>';
-          $result = $this->renderMenu($menu, $language, $this->getRequestParameter('assignedTo'), $page->getPageName(), $this->getRequestParameter('classAssigned'));
-
-          if ($idPage == $this->getRequestParameter('idPage')) $savedResult = $result;
-          if ($currentContent == null){
-            $currentContent = new W3sContent;
-            $aParams = array("idPage" => $idPage,
-                             "idLanguage" => $this->getRequestParameter('lang'),
-                             "idSlot" => $this->getRequestParameter('idSlot'),
-                             "idGroup" => $this->getRequestParameter('idGroup'),
-                             "idContent" => 0,
-                             "content" => $result,
-                             "contentType" => '5');
-          }
-          else{
-            $aParams = array("idPage" => $idPage,
-                             "idLanguage" => $this->getRequestParameter('lang'),
-                             "idSlot" => $this->getRequestParameter('idSlot'),
-                             "idContent" => $currentContent->getId(),
-                             "content" => $result,
-                             "contentType" => '5');
-          }
-          $currentContent->saveContent($aParams);
-          $oMenuElement->copyMenuFromContent($this->getRequestParameter('idContent'), $currentContent->getId());
-        endforeach;    
-      endforeach;
-    }
-    else{
-      $oPage = W3sPagePeer::retrieveByPk($this->getRequestParameter('idPage'));      
-      $savedResult = $this->renderMenu($menu, $language, $this->getRequestParameter('assignedTo'), $oPage->getPageName(), $this->getRequestParameter('classAssigned'));
-      
-      $contentValues = array("idPage"                 => $this->getRequestParameter('idPage'),
-                             "idLanguage"             => $this->getRequestParameter('lang'),
-                             "idSlot"                 => $this->getRequestParameter('idSlot'),
-                             "idContent"              => $this->getRequestParameter('idContent'),
-                             "contentType"            => '5',
-                             "content"                => $savedResult);
-      $content = new W3sContent();
-      if ($content->saveContent($contentValues) != 1){
-        $this->getResponse()->setStatusCode(404);
-        return sfView::NONE;
+      $i = 1;
+      $menuList = $this->getRequestParameter('w3s_mb_list');
+      foreach($menuList as $idMenu)
+      {
+        $menu = DbFinder::from('W3sMenuElement')->findPK($idMenu);
+        $menu->setPosition($i);
+        $menu->save();
+        $i++;
       }
-    }    
-    
-    if (!$bRollBack){ // Everything was fine so W3StudioCMS commits to database
-      $con->commit();
-      $result = 1;
+      
+      return sfView::NONE;
     }
-    else{             // Something was wrong so W3StudioCMS aborts the operation and restores to previous status
-      w3sPropelWorkaround::rollBack($con);
-      $result = 0;
+    else
+    {
+      $this->getResponse()->setStatusCode(404);
+      return $this->renderText('A required parameter misses.');
     }
-    
-    $this->clearCache($this->getRequestParameter('idContent'));
-   
-    return $this->renderText($savedResult);
   }*/
+
+
   
   /**
    * Clears the cache for the current content's modue
